2018/12/27
니폰이치 제출판
경고제거랑 함수,코멘트 정리

+++ Task_Game.cpp
@@ -119,7 +119,7 @@ namespace  Game
 		{
 			//šˆø‚«Œp‚¬ƒ^ƒXƒN‚Ì¶¬
 			
-			auto nextTask = Result::Object::Create(true,this->timeCnt,this->nowdi);
+			auto nextTask = Result::Object::Create(true,(int)this->timeCnt,this->nowdi);
 		}
 
 		return  true;
